State Display Upgrades

+++ Assets/Scripts/Moving Objects/Attributes/Health/Health.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 // basic interface for taking damage.
 public abstract class Health : MonoBehaviour
 {
-    public int health = 20;
+    public float health = 20;
     public OnBoard onBoard;
 
     public abstract void TakeDamage(int damage, Vector2 dir);
